need split horizon implementation, since in the server we only have the old requests module
